Adding flatten_dict util to simplify logic
