+++ src/fides/api/util/collection_util.py
@@ -1,3 +1,4 @@
+from collections import deque
 from functools import reduce
 from typing import Any, Callable, Dict, Iterable, List, Optional, TypeVar, Union
 
@@ -119,3 +120,119 @@ def extract_key_for_address(
     request_id_dataset, collection = full_request_id.split(":")
     dataset = request_id_dataset.split("__", number_of_leading_strings_to_exclude)[-1]
     return f"{dataset}:{collection}"
+
+
+def unflatten_dict(flat_dict: Dict[str, Any], separator: str = ".") -> Dict[str, Any]:
+    """
+    Converts a dictionary of paths/values into a nested dictionary
+
+    example:
+
+    {"A.B": "1", "A.C": "2"}
+
+    becomes
+
+    {
+        "A": {
+            "B": "1",
+            "C": "2"
+        }
+    }
+    """
+    output: Dict[Any, Any] = {}
+    queue = deque(flat_dict.items())
+
+    while queue:
+        path, value = queue.popleft()
+        keys = path.split(separator)
+        target = output
+        for i, current_key in enumerate(keys[:-1]):
+            next_key = keys[i + 1]
+            if next_key.isdigit():
+                target = target.setdefault(current_key, [])
+            else:
+                if isinstance(target, dict):
+                    target = target.setdefault(current_key, {})
+                elif isinstance(target, list):
+                    while len(target) <= int(current_key):
+                        target.append({})
+                    target = target[int(current_key)]
+        try:
+            if isinstance(target, list):
+                target.append(value)
+            else:
+                # If the value is a dictionary, add its components to the queue for processing
+                if isinstance(value, dict):
+                    target = target.setdefault(keys[-1], {})
+                    for inner_key, inner_value in value.items():
+                        new_key = f"{path}{separator}{inner_key}"
+                        queue.append((new_key, inner_value))
+                else:
+                    target[keys[-1]] = value
+        except TypeError as exc:
+            raise ValueError(
+                f"Error unflattening dictionary, conflicting levels detected: {exc}"
+            )
+    return output
+
+
+def flatten_dict(data: Any, prefix: str = "", separator: str = ".") -> Dict[str, Any]:
+    """
+    Recursively flatten a dictionary or list into a flat dictionary with dot-notation keys.
+    Handles nested dictionaries and arrays with proper indices.
+
+    example:
+
+    {
+        "A": {
+            "B": "1",
+            "C": "2"
+        },
+        "D": [
+            {"E": "3"},
+            {"E": "4"}
+        ]
+    }
+
+    becomes
+
+    {
+        "A.B": "1",
+        "A.C": "2",
+        "D.0.E": "3",
+        "D.1.E": "4"
+    }
+
+    Args:
+        data: The data to flatten (must be a dict or list)
+        prefix: The current key prefix (used in recursion)
+        separator: The separator to use between key segments (default: ".")
+
+    Returns:
+        A flattened dictionary with dot-notation keys
+
+    Raises:
+        FidesopsException: If input is not a dict or list
+    """
+    items = {}
+
+    if isinstance(data, dict):
+        for k, v in data.items():
+            new_key = f"{prefix}{separator}{k}" if prefix else k
+            if isinstance(v, (dict, list)):
+                items.update(flatten_dict(v, new_key, separator))
+            else:
+                items[new_key] = v
+    elif isinstance(data, list):
+        for i, v in enumerate(data):
+            new_key = f"{prefix}{separator}{i}"
+            if isinstance(v, (dict, list)):
+                items.update(flatten_dict(v, new_key, separator))
+            else:
+                items[new_key] = v
+    else:
+        raise ValueError(
+            f"Input to flatten_dict must be a dict or list, got {type(data).__name__}"
+        )
+
+    return items
