+++ tests/ops/util/test_collection_util.py
@@ -1,15 +1,19 @@
 from typing import Dict, List
 
+import pytest
 from immutables import Map
 from ordered_set import OrderedSet
 
+from fides.api.common_exceptions import FidesopsException
 from fides.api.util.collection_util import (
     append,
     filter_nonempty_values,
+    flatten_dict,
     make_immutable,
     make_mutable,
     merge_dicts,
     partition,
+    unflatten_dict,
 )
 
 
@@ -98,3 +102,179 @@ class TestMutabilityConversion:
         str_obj = "hello"
         assert make_mutable(int_obj) == 42
         assert make_mutable(str_obj) == "hello"
+
+
+@pytest.mark.unit
+class TestUnflattenDict:
+    def test_empty_dict(self):
+        assert unflatten_dict({}) == {}
+
+    def test_empty_dict_value(self):
+        assert unflatten_dict({"A": {}}) == {"A": {}}
+
+    def test_unflattened_dict(self):
+        assert unflatten_dict({"A": "1"}) == {"A": "1"}
+
+    def test_same_level(self):
+        assert unflatten_dict({"A.B": "1", "A.C": "2"}) == {"A": {"B": "1", "C": "2"}}
+
+    def test_mixed_levels(self):
+        assert unflatten_dict(
+            {
+                "A": "1",
+                "B.C": "2",
+                "B.D": "3",
+            }
+        ) == {
+            "A": "1",
+            "B": {"C": "2", "D": "3"},
+        }
+
+    def test_long_path(self):
+        assert unflatten_dict({"A.B.C.D.E.F.G": "1"}) == {
+            "A": {"B": {"C": {"D": {"E": {"F": {"G": "1"}}}}}}
+        }
+
+    def test_single_item_array(self):
+        assert unflatten_dict({"A.0.B": "C"}) == {"A": [{"B": "C"}]}
+
+    def test_multi_item_array(self):
+        assert unflatten_dict({"A.0.B": "C", "A.1.D": "E"}) == {
+            "A": [{"B": "C"}, {"D": "E"}]
+        }
+
+    def test_multi_value_array(self):
+        assert unflatten_dict(
+            {"A.0.B": "C", "A.0.D": "E", "A.1.F": "G", "A.1.H": "I"}
+        ) == {"A": [{"B": "C", "D": "E"}, {"F": "G", "H": "I"}]}
+
+    def test_array_with_scalar_value(self):
+        assert unflatten_dict({"A.0": "B"}) == {"A": ["B"]}
+
+    def test_array_with_scalar_values(self):
+        assert unflatten_dict({"A.0": "B", "A.1": "C"}) == {"A": ["B", "C"]}
+
+    def test_overwrite_existing_values(self):
+        assert unflatten_dict({"A.B": 1, "A.B": 2}) == {"A": {"B": 2}}
+
+    def test_conflicting_types(self):
+        with pytest.raises(FidesopsException):
+            unflatten_dict({"A.B": 1, "A": 2, "A.C": 3})
+
+    def test_mixed_types_in_array(self):
+        assert unflatten_dict({"A.0": "B", "A.1.C": "D"}) == {"A": ["B", {"C": "D"}]}
+
+    def test_data_not_completely_flattened(self):
+        assert unflatten_dict({"A.B.C": 1, "A": {"B.D": 2}}) == {
+            "A": {"B": {"C": 1, "D": 2}}
+        }
+
+    def test_response_with_object_fields_specified(self):
+        assert unflatten_dict(
+            {
+                "address.email": "2a3aaa22b2ccce15ef7a1e94ee@email.com",
+                "address.name": "MASKED",
+                "metadata": {"age": "24", "place": "Bedrock"},
+                "return_path": "",
+                "substitution_data": {
+                    "favorite_color": "SparkPost Orange",
+                    "job": "Software Engineer",
+                },
+                "tags": ["greeting", "prehistoric", "fred", "flintstone"],
+            }
+        ) == {
+            "address": {
+                "email": "2a3aaa22b2ccce15ef7a1e94ee@email.com",
+                "name": "MASKED",
+            },
+            "metadata": {"age": "24", "place": "Bedrock"},
+            "return_path": "",
+            "substitution_data": {
+                "favorite_color": "SparkPost Orange",
+                "job": "Software Engineer",
+            },
+            "tags": ["greeting", "prehistoric", "fred", "flintstone"],
+        }
+
+    def test_none_separator(self):
+        with pytest.raises(IndexError):
+            unflatten_dict({"": "1"}, separator=None)
+
+
+@pytest.mark.unit
+class TestFlattenDict:
+    def test_empty_dict(self):
+        assert flatten_dict({}) == {}
+
+    def test_scalar_value(self):
+        with pytest.raises(ValueError) as exc:
+            flatten_dict(42)
+        assert "Input to flatten_dict must be a dict or list" in str(exc)
+
+    def test_unnested_dict(self):
+        assert flatten_dict({"A": "1", "B": "2"}) == {"A": "1", "B": "2"}
+
+    def test_nested_dict(self):
+        assert flatten_dict({"A": {"B": "1", "C": "2"}}) == {"A.B": "1", "A.C": "2"}
+
+    def test_deep_nesting(self):
+        assert flatten_dict({"A": {"B": {"C": {"D": {"E": {"F": {"G": "1"}}}}}}}) == {
+            "A.B.C.D.E.F.G": "1"
+        }
+
+    def test_list_values(self):
+        assert flatten_dict({"A": ["B", "C"]}) == {"A.0": "B", "A.1": "C"}
+
+    def test_list_nested_dict(self):
+        assert flatten_dict({"A": [{"B": "C"}, {"D": "E"}]}) == {
+            "A.0.B": "C",
+            "A.1.D": "E",
+        }
+
+    def test_mixed_nested_structures(self):
+        data = {"A": {"B": "1", "C": "2"}, "D": [{"E": "3"}, {"E": "4"}]}
+        expected = {"A.B": "1", "A.C": "2", "D.0.E": "3", "D.1.E": "4"}
+        assert flatten_dict(data) == expected
+
+    def test_empty_nested_structures(self):
+        # Empty dictionaries and arrays don't contribute any keys to the flattened result
+        assert flatten_dict({"A": {}, "B": []}) == {}
+
+        # Mixed with non-empty values
+        assert flatten_dict({"A": {}, "B": [], "C": "value"}) == {"C": "value"}
+
+    def test_none_values(self):
+        assert flatten_dict({"A": None, "B": {"C": None}}) == {"A": None, "B.C": None}
+
+    def test_custom_separator(self):
+        assert flatten_dict({"A": {"B": "1"}}, separator="_") == {"A_B": "1"}
+
+    def test_with_prefix(self):
+        assert flatten_dict({"A": "1"}, prefix="prefix") == {"prefix.A": "1"}
+
+    def test_with_prefix_and_custom_separator(self):
+        assert flatten_dict({"A": "1"}, prefix="prefix", separator="_") == {
+            "prefix_A": "1"
+        }
+
+    def test_complex_example(self):
+        data = {
+            "address": {
+                "email": "test@example.com",
+                "name": "Test User",
+            },
+            "metadata": {"age": "24", "place": "Anywhere"},
+            "empty_field": "",
+            "tags": ["greeting", "test", "example"],
+        }
+        expected = {
+            "address.email": "test@example.com",
+            "address.name": "Test User",
+            "metadata.age": "24",
+            "metadata.place": "Anywhere",
+            "empty_field": "",
+            "tags.0": "greeting",
+            "tags.1": "test",
+            "tags.2": "example",
+        }
+        assert flatten_dict(data) == expected
