+++ tests/ops/util/test_saas_util.py
@@ -1,6 +1,5 @@
 import pytest
 
-from fides.api.common_exceptions import FidesopsException
 from fides.api.graph.config import (
     Collection,
     FieldAddress,
@@ -9,11 +8,11 @@ from fides.api.graph.config import (
     ObjectField,
     ScalarField,
 )
+from fides.api.util.collection_util import unflatten_dict
 from fides.api.util.saas_util import (
     assign_placeholders,
     merge_datasets,
     replace_version,
-    unflatten_dict,
 )
 
 
@@ -504,99 +503,6 @@ class TestAssignPlaceholders:
             )
             == '{"subscriber_ids": []}'
         )
-
-
-@pytest.mark.unit_saas
-class TestUnflattenDict:
-    def test_empty_dict(self):
-        assert unflatten_dict({}) == {}
-
-    def test_empty_dict_value(self):
-        assert unflatten_dict({"A": {}}) == {"A": {}}
-
-    def test_unflattened_dict(self):
-        assert unflatten_dict({"A": "1"}) == {"A": "1"}
-
-    def test_same_level(self):
-        assert unflatten_dict({"A.B": "1", "A.C": "2"}) == {"A": {"B": "1", "C": "2"}}
-
-    def test_mixed_levels(self):
-        assert unflatten_dict(
-            {
-                "A": "1",
-                "B.C": "2",
-                "B.D": "3",
-            }
-        ) == {
-            "A": "1",
-            "B": {"C": "2", "D": "3"},
-        }
-
-    def test_long_path(self):
-        assert unflatten_dict({"A.B.C.D.E.F.G": "1"}) == {
-            "A": {"B": {"C": {"D": {"E": {"F": {"G": "1"}}}}}}
-        }
-
-    def test_single_item_array(self):
-        assert unflatten_dict({"A.0.B": "C"}) == {"A": [{"B": "C"}]}
-
-    def test_multi_item_array(self):
-        assert unflatten_dict({"A.0.B": "C", "A.1.D": "E"}) == {
-            "A": [{"B": "C"}, {"D": "E"}]
-        }
-
-    def test_multi_value_array(self):
-        assert unflatten_dict(
-            {"A.0.B": "C", "A.0.D": "E", "A.1.F": "G", "A.1.H": "I"}
-        ) == {"A": [{"B": "C", "D": "E"}, {"F": "G", "H": "I"}]}
-
-    def test_array_with_scalar_value(self):
-        assert unflatten_dict({"A.0": "B"}) == {"A": ["B"]}
-
-    def test_array_with_scalar_values(self):
-        assert unflatten_dict({"A.0": "B", "A.1": "C"}) == {"A": ["B", "C"]}
-
-    def test_overwrite_existing_values(self):
-        assert unflatten_dict({"A.B": 1, "A.B": 2}) == {"A": {"B": 2}}
-
-    def test_conflicting_types(self):
-        with pytest.raises(FidesopsException):
-            unflatten_dict({"A.B": 1, "A": 2, "A.C": 3})
-
-    def test_mixed_types_in_array(self):
-        assert unflatten_dict({"A.0": "B", "A.1.C": "D"}) == {"A": ["B", {"C": "D"}]}
-
-    def test_data_not_completely_flattened(self):
-        assert unflatten_dict({"A.B.C": 1, "A": {"B.D": 2}}) == {
-            "A": {"B": {"C": 1, "D": 2}}
-        }
-
-    def test_response_with_object_fields_specified(self):
-        assert unflatten_dict(
-            {
-                "address.email": "2a3aaa22b2ccce15ef7a1e94ee@email.com",
-                "address.name": "MASKED",
-                "metadata": {"age": "24", "place": "Bedrock"},
-                "return_path": "",
-                "substitution_data": {
-                    "favorite_color": "SparkPost Orange",
-                    "job": "Software Engineer",
-                },
-                "tags": ["greeting", "prehistoric", "fred", "flintstone"],
-            }
-        ) == {
-            "address": {
-                "email": "2a3aaa22b2ccce15ef7a1e94ee@email.com",
-                "name": "MASKED",
-            },
-            "metadata": {"age": "24", "place": "Bedrock"},
-            "return_path": "",
-            "substitution_data": {
-                "favorite_color": "SparkPost Orange",
-                "job": "Software Engineer",
-            },
-            "tags": ["greeting", "prehistoric", "fred", "flintstone"],
-        }
 
     def test_none_separator(self):
         with pytest.raises(IndexError):
